+++ adapters/AdColony/CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Next Version
 - Updated the adapter to use the `didRewardUser` API.
 - Fixed an issue where UI related API methods are being called on a background thread.
+- Removed use of deprecated gender, birthday and location mediation APIs.
 - Now requires Google Mobile Ads SDK version 9.8.0 or higher.
 
 #### [Version 4.9.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/adcolony/AdColonyAdapter-4.9.0.0.zip)